Solution

+++ app/create_file.py
@@ -4,10 +4,12 @@ from datetime import datetime
 
 
 class FlagError(Exception):
+    """An exception that occurs when there is an error related to flags."""
     pass
 
 
 class ArgumentError(Exception):
+    """Exception that occurs when there is an error related to command line arguments."""
     pass
 
 
@@ -19,13 +21,16 @@ def if_flag(argument: str) -> bool:
 def main(args: list[str]) -> None:
     available_flags = ["-d", "-f"]
     flags = [flag for flag in args if if_flag(flag)]
+
     if not any([True if flag in available_flags else False for flag in flags]):
         raise FlagError("You must enter at least one flag")
     if not all([True if flag in available_flags else False for flag in flags]):
         raise FlagError("Invalid flags entered")
+
     index_dict = {}
     for flag in flags:
         index_dict[flag] = args.index(flag)
+
     flags_args = {}
     for flag in flags:
         flags_args[flag] = []
@@ -33,7 +38,9 @@ def main(args: list[str]) -> None:
             if if_flag(args[i]):
                 break
             flags_args[flag].append(args[i])
+
     base_path = ""
+
     if "-d" in flags:
         if len(flags_args["-d"]) == 0:
             raise ArgumentError(
@@ -41,23 +48,35 @@ def main(args: list[str]) -> None:
             )
         base_path = os.path.join(*flags_args["-d"])
         os.makedirs(base_path, exist_ok=True)
+
     if "-f" in flags:
         if len(flags_args["-f"]) == 0:
             raise ArgumentError(
                 "At least one argument is required for the -f flag"
             )
+
         file_path = os.path.join(base_path, flags_args["-f"][0])
+
         input_list = []
         while True:
             line = input("Enter content line: ")
             if line == "stop":
                 break
             input_list.append(line)
-        with open(file_path, "w") as file:
+
+        if not os.path.isfile(file_path):
+            with open(file_path, "w") as file:
+                current_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+                file.write(current_time + "\n")
+                for index, line in enumerate(input_list):
+                    file.write(f"{index + 1} {line}\n")
+        with open(file_path, "a") as file:
             current_time = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+            file.write("\n")
             file.write(current_time + "\n")
             for index, line in enumerate(input_list):
                 file.write(f"{index + 1} {line}\n")
 
 
-main(sys.argv)
+if __name__ == "__main__":
+    main(sys.argv)
